+++ src/test/java/com/dejvo/Shop/db/crudservice/ProductInterfaceTest.java
@@ -10,6 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.Instant;
 
@@ -22,7 +23,7 @@ class ProductInterfaceTest {
 
     @Test
     public void createProduct(){
-        Product product= new Product(1L,"Lepidlo","Biele lepidlo lacne",0.55,10, Timestamp.from(Instant.now()));
+        Product product= new Product(1L,"Lepidlo","Biele lepidlo lacne",BigDecimal.valueOf(0.55),10, Timestamp.from(Instant.now()));
         assertNotNull(productInterface.createProduct(product));
     }
 
@@ -43,12 +44,12 @@ class ProductInterfaceTest {
     }
     @Test
     public void updateProductById(){
-        UpdateProductRequest request= new UpdateProductRequest("Taska Rapotaska","Trocha ina taska ale dobra",0.69,25);
+        UpdateProductRequest request= new UpdateProductRequest("Taska Rapotaska","Trocha ina taska ale dobra", BigDecimal.valueOf(0.69),25);
         assertEquals(1,productInterface.updateProduct(request,3));
     }
     @Test
     public void updateProductWithNonExistingId(){
-        UpdateProductRequest request= new UpdateProductRequest("Taska Rapotaska","Trocha ina taska ale dobra",0.69,25);
+        UpdateProductRequest request= new UpdateProductRequest("Taska Rapotaska","Trocha ina taska ale dobra",BigDecimal.valueOf(0.69),25);
         assertEquals(0,productInterface.updateProduct(request,355));
     }
 }
